# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/listings": {
      "get": {
        "operationId": "AdminController_adminListings",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/analytics/events": {
      "post": {
        "operationId": "AnalyticsController_ingest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestEventsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
...
    "/api/analytics/admin/guest-member-split": {
      "get": {
        "operationId": "AnalyticsController_split",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/admin/top-searches": {
      "get": {
        "operationId": "AnalyticsController_topSearches",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "zeroOnly",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/admin/funnel/purchase": {
      "get": {
        "operationId": "AnalyticsController_purchaseFunnel",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/admin/events": {
      "get": {
        "operationId": "AnalyticsController_explore",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/admin/timeseries": {
      "get": {
        "operationId": "AnalyticsController_timeseries",
        "parameters": [
          {
            "name": "interval",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minutes",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "event",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/admin/listings": {
      "get": {
        "operationId": "AnalyticsController_listingStats",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/admin/order-outcomes": {
      "get": {
        "operationId": "AnalyticsController_orderOutcomes",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/api/analytics/admin/funnel/purchase-by-tier": {
      "get": {
        "operationId": "AnalyticsController_purchaseFunnelByTier",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
...
    "/api/analytics/admin/zero-result-trend": {
      "get": {
        "operationId": "AnalyticsController_zeroResultTrend",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
...
  },
  "info": {
    "title": "Certifine API",
    "description": "C2C marketplace with per-category third-party authentication. Generated from the running app — do not hand-edit, and do not hand-edit the client generated from it.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "IngestEventsDto": {
        "type": "object",
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "events"
        ]
      }
    }
  }
}